utils/requirements.txt
freetype-py
Pillow
pypng
